+++ src/app/freelancer/project/[project_id]/page.tsx
@@ -88,7 +88,7 @@ export default function Dashboard() {
           <Breadcrumb
             items={[
               { label: 'Freelancer', link: '/dashboard/freelancer' },
-              { label: 'Project', link: '/market/freelancer' },
+              { label: 'Project', link: '/freelancer/market' },
               { label: project_id, link: '#' },
             ]}
           />
