+++ src/app/freelancer/talent/page.tsx
