+++ src/config/menuItems/freelancer/interviewMenuItems.tsx
@@ -4,22 +4,22 @@ import { MenuItem } from '@/components/menu/sidebarMenu';
 
 export const menuItemsTop: MenuItem[] = [
   {
-    href: '/dashboard/freelancer/interview',
+    href: '/freelancer/interview',
     icon: <Boxes className="h-5 w-5" />,
     label: 'Dehix',
   },
   {
-    href: '/dashboard/freelancer/interview/profile',
+    href: '/freelancer/interview/profile',
     icon: <Users2 className="h-5 w-5" />,
     label: 'Profile',
   },
   {
-    href: '/dashboard/freelancer/interview/current',
+    href: '/freelancer/interview/current',
     icon: <ListVideo className="h-5 w-5" />,
     label: 'Current',
   },
   {
-    href: '/dashboard/freelancer/interview/history',
+    href: '/freelancer/interview/history',
     icon: <History className="h-5 w-5" />,
     label: 'History',
   },
