+++ src/config/menuItems/freelancer/oracleMenuItems.tsx
@@ -16,22 +16,22 @@ export const menuItemsTop: MenuItem[] = [
     label: 'Dehix',
   },
   {
-    href: '/dashboard/freelancer/oracleDashboard/businessVerification',
+    href: '/freelancer/oracleDashboard/businessVerification',
     icon: <Briefcase className="h-5 w-5" />,
     label: 'Business Verification',
   },
   {
-    href: '/dashboard/freelancer/oracleDashboard/workExpVerification',
+    href: '/freelancer/oracleDashboard/workExpVerification',
     icon: <User className="h-5 w-5" />,
     label: 'Experience Verification',
   },
   {
-    href: '/dashboard/freelancer/oracleDashboard/projectVerification',
+    href: '/freelancer/oracleDashboard/projectVerification',
     icon: <Package className="h-5 w-5" />,
     label: 'Project Verification',
   },
   {
-    href: '/dashboard/freelancer/oracleDashboard/educationVerification',
+    href: '/freelancer/oracleDashboard/educationVerification',
     icon: <BookOpen className="h-5 w-5" />,
     label: 'Education Verification',
   },
