+++ src/config/menuItems/freelancer/projectMenuItems.tsx
@@ -16,22 +16,22 @@ export const menuItemsTop: MenuItem[] = [
     label: 'Dehix',
   },
   {
-    href: '/project/current',
+    href: '/freelancer/project/current',
     icon: <FolderDot className="h-5 w-5" />,
     label: 'Current Projects',
   },
   {
-    href: '/project/applied',
+    href: '/freelancer/project/applied',
     icon: <Pointer className="h-5 w-5" />,
     label: 'Under Verification',
   },
   {
-    href: '/project/completed',
+    href: '/freelancer/project/completed',
     icon: <FileCheck className="h-5 w-5" />,
     label: 'Completed Projects',
   },
   {
-    href: '/project/rejected',
+    href: '/freelancer/project/rejected',
     icon: <CircleX className="h-5 w-5" />,
     label: 'Rejected Verification',
   },
